Rename changelog scripts to changelog_ prefix; add changelog-deploy recipe
Rename consolidate_changelog.py -> changelog_consolidate.py,
trigger_changelog_consolidation.py -> changelog_consolidation_trigger.py, and
setup_changelog_agent.sh -> changelog_deploy.sh so the changelog tooling sorts
together. Update all imports, docstrings, and the consolidation prompt. Add a
changelog-deploy justfile recipe wrapping scripts/changelog_deploy.sh.
