+++ scripts/changelog_consolidation_trigger_test.py
@@ -2,11 +2,11 @@ import subprocess
 import sys
 from pathlib import Path
 
-from scripts.trigger_changelog_consolidation import PROVIDER
-from scripts.trigger_changelog_consolidation import _ENABLED_PLUGINS
-from scripts.trigger_changelog_consolidation import disable_plugin_args
+from scripts.changelog_consolidation_trigger import PROVIDER
+from scripts.changelog_consolidation_trigger import _ENABLED_PLUGINS
+from scripts.changelog_consolidation_trigger import disable_plugin_args
 
-_SCRIPT_PATH = Path(__file__).resolve().parent / "trigger_changelog_consolidation.py"
+_SCRIPT_PATH = Path(__file__).resolve().parent / "changelog_consolidation_trigger.py"
 
 
 def test_disable_plugin_args_returns_paired_flags() -> None:
@@ -24,7 +24,7 @@ def test_disable_plugin_args_returns_paired_flags() -> None:
 
 
 def test_cli_print_disable_plugin_args_matches_helper() -> None:
-    """The CLI flag is the integration point with setup_changelog_agent.sh;
+    """The CLI flag is the integration point with changelog_deploy.sh;
     its output must match the in-process helper.
     """
     result = subprocess.run(
@@ -38,7 +38,7 @@ def test_cli_print_disable_plugin_args_matches_helper() -> None:
 
 
 def test_cli_print_provider_matches_constant() -> None:
-    """The CLI flag is the integration point with setup_changelog_agent.sh
+    """The CLI flag is the integration point with changelog_deploy.sh
     (which reads it to set $PROVIDER); its output must match the in-process
     constant so the deploy and release.py's printed on-demand command can't
     drift.
